# src/ClientApiSdkGen/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SMEIoT API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "https://localhost:5001"
    },
    {
      "url": "https://smeiot.se"
    }
  ],
  "paths": {
...
    "/api/devices/bootstrap": {
      "post": {
        "tags": [
          "Devices"
        ],
        "requestBody": {
          "content": {
            "application/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/DeviceBootstrapConfigBindingModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicDeviceApiModel"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{name}": {
      "put": {
        "tags": [
          "Devices"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/DeviceConfigBindingModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicDeviceApiModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Devices"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDetailsApiModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{name}/basic": {
      "get": {
        "tags": [
          "Devices"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicDeviceApiModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{name}/sensor_candidates": {
      "get": {
        "tags": [
          "Devices"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorCandidatesApiModel"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/devices": {
      "get": {
        "tags": [
          "Devices"
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicDeviceApiModelList"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/sensors/{name}/users": {
      "post": {
        "tags": [
          "SensorAssignments"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/AssignUserSensorBindingModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorAssignmentApiModel"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "SensorAssignments"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSensorIAsyncEnumerable"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/sensors/{name}/users/{userName}": {
      "delete": {
        "tags": [
          "SensorAssignments"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Success"
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/sensors": {
      "get": {
        "tags": [
          "Sensors"
        ],
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensorDetailsApiModelList"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Sensors"
        ],
        "requestBody": {
          "content": {
            "application/json; v=1.0": {
              "schema": {
                "$ref": "#/components/schemas/SensorLocatorBindingModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicSensorApiModel"
                }
              }
            }
          },
          "422": {
            "description": "Client Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "extensions": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "additionalProperties": false
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
...
      "DeviceBootstrapConfigBindingModel": {
        "required": [
          "key",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "key": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "DeviceAuthenticationType": {
        "enum": [
          0
        ],
        "type": "integer",
        "format": "int32"
      },
      "BasicDeviceApiModel": {
        "required": [
          "authenticationType",
          "connected",
          "connectedAt",
          "createdAt",
          "lastMessageAt",
          "name",
          "preSharedKey",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "readOnly": true
          },
          "createdAt": {
            "type": "string",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "readOnly": true
          },
          "authenticationType": {
            "$ref": "#/components/schemas/DeviceAuthenticationType"
          },
          "preSharedKey": {
            "type": "string",
            "readOnly": true
          },
          "connected": {
            "type": "boolean",
            "readOnly": true
          },
          "connectedAt": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "lastMessageAt": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DeviceConfigBindingModel": {
        "required": [
          "key"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "BasicSensorApiModel": {
        "required": [
          "deviceName",
          "sensorName"
        ],
        "type": "object",
        "properties": {
          "sensorName": {
            "type": "string"
          },
          "deviceName": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "DeviceDetailsApiModel": {
        "required": [
          "authenticationType",
          "connected",
          "connectedAt",
          "createdAt",
          "lastMessageAt",
          "name",
          "preSharedKey",
          "sensors",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "sensors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BasicSensorApiModel"
            },
            "readOnly": true
          },
          "name": {
            "type": "string",
            "readOnly": true
          },
          "createdAt": {
            "type": "string",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "readOnly": true
          },
          "authenticationType": {
            "$ref": "#/components/schemas/DeviceAuthenticationType"
          },
          "preSharedKey": {
            "type": "string",
            "readOnly": true
          },
          "connected": {
            "type": "boolean",
            "readOnly": true
          },
          "connectedAt": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "lastMessageAt": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "SensorCandidatesApiModel": {
        "required": [
          "names"
        ],
        "type": "object",
        "properties": {
          "names": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
...
      "BasicDeviceApiModelList": {
        "required": [
          "devices"
        ],
        "type": "object",
        "properties": {
          "devices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BasicDeviceApiModel"
            }
          }
        },
        "additionalProperties": false
      },
      "AssignUserSensorBindingModel": {
        "required": [
          "userName"
        ],
        "type": "object",
        "properties": {
          "userName": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "SensorAssignmentApiModel": {
        "required": [
          "assignmentCount",
          "name",
          "userName"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "userName": {
            "type": "string"
          },
          "assignmentCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "UserSensorIAsyncEnumerable": {
        "type": "object",
        "additionalProperties": false
      },
      "NumberTimeSeriesApiModel": {
        "required": [
          "createdAt",
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "SensorDetailsApiModel": {
        "required": [
          "data",
          "deviceName",
          "sensorName",
          "startedAt"
        ],
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NumberTimeSeriesApiModel"
            }
          },
          "startedAt": {
            "type": "string"
          },
          "sensorName": {
            "type": "string"
          },
          "deviceName": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "SensorDetailsApiModelList": {
        "required": [
          "sensors"
        ],
        "type": "object",
        "properties": {
          "sensors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SensorDetailsApiModel"
            }
          }
        },
        "additionalProperties": false
      },
      "SensorLocatorBindingModel": {
        "required": [
          "deviceName",
          "name"
        ],
        "type": "object",
        "properties": {
          "deviceName": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}